--- python/src/coreai_models/models/gpu/__init__.py
@@ -1,4 +1,0 @@
-# Copyright 2026 Apple Inc.
-#
-# Use of this source code is governed by a BSD-3-clause license that can
-# be found in the LICENSE file or at https://opensource.org/licenses/BSD-3-Clause
