+++ python/src/coreai_models/models/macos/qwen3_vl.py
